support for multiple data sources added and output improved
- multiple data sources can now be used simultaniously
- data sources can be prioritized
- output design further improved

+++ bitcoin.bat
@@ -1,5 +1,6 @@
 @echo off
+::end old cgminer if running
+tasklist /FI "IMAGENAME eq cgminer.exe" 2>NUL | find /I /N "cgminer.exe">NUL
+if "%ERRORLEVEL%"=="0" taskkill /IM cgminer.exe /F
 ::start new cgminer
-taskkill /IM cgminer.exe /F
-::start new cgminer
-start cmd.exe /C "cgminer\cgminer.exe -o mint.bitminter.com:8332 -u l05443.rws -p rs1372 -I 9 --gpu-reorder -k diablo --worksize 256 --lookup-gap 2 --auto-fan --api-network --api-listen --api-allow W:127.0.0.1"
+start cmd.exe /C "cgminer\cgminer.exe -o mint.bitminter.com:8332 -u USER -p PASSWORD -I 9 --gpu-reorder -k diablo --worksize 256 --lookup-gap 2 --auto-fan"
